requirements_all.txt
# Package==Version
# -------------------------==--------------
absl-py==1.0.0
aiofiles==23.1.0
aiohttp==3.8.4
aiosignal==1.3.1
altair==5.0.1
antlr4-python3-runtime==4.9.3
anyio==3.7.1
apex==0.1
appdirs==1.4.4
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
arrow==1.2.3
asttokens==2.2.1
astunparse==1.6.3
async-lru==2.0.4
async-timeout==4.0.2
attrs==22.2.0
av==10.0.0
Babel==2.12.1
backcall==0.2.0
backports.zoneinfo==0.2.1
beautifulsoup4==4.11.2
bitsandbytes==0.37.0
bleach==6.0.0
blinker==1.6.2
blis==0.7.10
boto3==1.22.13
botocore==1.25.13
braceexpand==0.1.7
cachetools==5.0.0
catalogue==2.0.9
cchardet==2.1.7
certifi==2021.10.8
cffi==1.15.1
cfgv==3.3.1
chardet==5.1.0
charset-normalizer==2.0.12
click==8.1.3
cmake==3.27.0
comm==0.1.2
confection==0.1.0
contexttimer==0.3.3
contourpy==1.0.7
cycler==0.11.0
cymem==2.0.7
datasets==2.9.0
debugpy==1.6.6
decorator==5.1.1
decord==0.6.0
deepspeed==0.9.0+970d827f
defusedxml==0.7.1
dill==0.3.6
distlib==0.3.7
docker-pycreds==0.4.0
einops==0.4.1
evaluate==0.4.0
exceptiongroup==1.1.2
executing==1.2.0
expecttest==0.1.3
fastapi==0.100.1
fastjsonschema==2.16.2
ffmpy==0.3.1
filelock==3.12.2
fonttools==4.38.0
fqdn==1.5.1
frozenlist==1.3.3
fschat==0.2.0
fsspec==2023.1.0
ftfy==6.1.1
future==0.18.2
git-lfs==1.6
gitdb==4.0.10
GitPython==3.1.31
google-auth==2.6.6
google-auth-oauthlib==0.4.6
gradio==3.23.0
grpcio==1.46.1
h11==0.14.0
hjson==3.1.0
httpcore==0.17.3
httpx==0.24.1
huggingface-hub==0.16.4
identify==2.5.26
idna==3.3
imageio==2.31.1
importlib-metadata==4.11.3
importlib-resources==5.12.0
ipykernel==6.21.2
ipython==8.10.0
isoduration==20.11.0
iterative-stratification==0.1.7
jedi==0.18.2
jieba==0.42.1
Jinja2==3.1.2
jmespath==1.0.0
joblib==1.2.0
json-lines==0.5.0
json5==0.9.14
jsonpointer==2.4
jsonschema==4.18.6
jsonschema-specifications==2023.7.1
jupyter_client==8.0.3
jupyter_core==5.2.0
jupyter-events==0.7.0
jupyter-lsp==2.2.0
jupyter_server==2.7.0
jupyter_server_terminals==0.4.4
jupyterlab==4.0.3
jupyterlab-pygments==0.2.2
jupyterlab_server==2.24.0
kaggle==1.5.16
kiwisolver==1.4.4
langcodes==3.3.0
lazy_loader==0.3
linkify-it-py==2.0.2
lit==16.0.6
llvmlite==0.40.1
Markdown==3.3.7
markdown-it-py==2.2.0
markdown2==2.4.10
MarkupSafe==2.1.2
matplotlib==3.7.0
matplotlib-inline==0.1.6
mdit-py-plugins==0.3.3
mdurl==0.1.2
mistune==2.0.5
mpmath==1.3.0
multidict==6.0.4
multiprocess==0.70.14
murmurhash==1.0.9
natsort==8.2.0
nbclient==0.7.2
nbconvert==7.2.9
nbformat==5.7.3
nest-asyncio==1.5.6
networkx==3.1
ninja==1.11.1
nlpcda==2.5.8
nltk==3.8.1
nodeenv==1.8.0
notebook==7.0.1
notebook_shim==0.2.3
numba==0.57.1
numpy==1.22.3
nvidia-cublas-cu11==11.10.3.66
nvidia-cuda-cupti-cu11==11.7.101
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cudnn-cu11==8.5.0.96
nvidia-cufft-cu11==10.9.0.58
nvidia-curand-cu11==10.2.10.91
nvidia-cusolver-cu11==11.4.0.1
nvidia-cusparse-cu11==11.7.4.91
nvidia-nccl-cu11==2.14.3
nvidia-nvtx-cu11==11.7.91
oauthlib==3.2.0
openai==0.27.0
opencv-contrib-python==4.5.5.64
opencv-python==4.7.0.72
opencv-python-headless==4.5.5.64
opendatasets==0.1.22
orjson==3.9.2
overrides==7.3.1
packaging==23.0
pandas==1.4.2
pandocfilters==1.5.0
parso==0.8.3
pathlib==1.0.1
pathtools==0.1.2
pathy==0.10.2
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.1.0
pip==23.2.1
pkgutil_resolve_name==1.3.10
platformdirs==3.10.0
plotly==5.15.0
portalocker==2.7.0
pre-commit==3.3.3
preshed==3.0.8
prettytable==3.6.0
prometheus-client==0.17.1
prompt-toolkit==3.0.36
protobuf==3.20.1
psutil==5.9.4
ptyprocess==0.7.0
pure-eval==0.2.2
py-cpuinfo==9.0.0
pyarrow==8.0.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pybind11==2.9.2
pycocotools==2.0.6
pycparser==2.21
pydantic==1.10.12
pydeck==0.8.0
pyDeprecate==0.3.1
pydub==0.25.1
Pygments==2.14.0
Pympler==1.0.1
pynndescent==0.5.10
pyparsing==3.0.9
pyrsistent==0.19.3
python-dateutil==2.8.2
python-json-logger==2.0.7
python-magic==0.4.27
python-multipart==0.0.6
python-slugify==8.0.1
pytorch-lightning==1.5.10
pytz==2022.1
pytz-deprecation-shim==0.1.0.post0
PyWavelets==1.4.1
PyYAML==6.0
pyzmq==25.0.0
referencing==0.30.0
regex==2022.10.31
requests==2.31.0
requests-oauthlib==1.3.1
responses==0.18.0
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rich==13.5.2
rpds-py==0.9.2
rsa==4.8
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.6
runipy==0.1.5
s3transfer==0.5.2
safetensors==0.3.1
scikit-image==0.21.0
scikit-learn==1.2.2
scipy==1.10.1
semantic-version==2.10.0
Send2Trash==1.8.2
sentencepiece==0.1.97
sentry-sdk==1.15.0
setproctitle==1.3.2
setuptools==59.5.0
six==1.16.0
smart-open==6.3.0
smmap==5.0.0
sniffio==1.3.0
soupsieve==2.4
sox==1.4.1
spacy-legacy==3.0.12
spacy-loggers==1.0.4
srsly==2.4.7
stack-data==0.6.2
starlette==0.27.0
streamlit==1.25.0
svgwrite==1.4.3
sympy==1.12
tenacity==8.2.2
tensorboard==2.9.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorboardX==2.5
termcolor==1.1.0
terminado==0.17.1
text-unidecode==1.3
thinc==8.1.10
threadpoolctl==3.1.0
tifffile==2023.7.10
timm==0.4.12
tinycss2==1.2.1
tokenizers==0.13.2
toml==0.10.2
tomli==2.0.1
toolz==0.12.0
torch==2.0.1
torchmetrics==0.11.1
tornado==6.2
tqdm==4.64.1
traitlets==5.9.0
triton==2.0.0
typer==0.7.0
types-dataclasses==0.6.5
typing_extensions==4.7.1
tzdata==2023.3
tzlocal==4.3.1
uc-micro-py==1.0.2
umap-learn==0.5.3
uri-template==1.3.0
urllib3==1.26.14
uvicorn==0.23.2
validators==0.20.0
virtualenv==20.24.2
wandb==0.13.10
wasabi==1.1.2
watchdog==3.0.0
wavedrom==2.0.3.post3
wcwidth==0.2.5
webcolors==1.13
webencodings==0.5.1
websocket-client==1.6.1
websockets==11.0.3
Werkzeug==2.1.2
wheel==0.37.1
xxhash==3.2.0
yarl==1.8.2
zipp==3.14.0
